The <Dataset> element has been added.  A <Dataset> is a collection of <Image>s that are processed together.  The <Project> element is a collection of <Dataset>s , and is no longer directly associated with images.  <Image> has zero or more references to <Dataset> through <DatasetRef>.  <Dataset> in turn has zero or more references to <Project> through <ProjectRef>.  <Project>, <Dataset>, and <Image> each have their own separate <Description>, <Experimenter> and <Group>.
An instrument may now be composed of multiple <LightSource>s,
<Filter>s, <Objective>s, <Camera>s, and <OTF>s.  There is still only
one <Microscope> element. All of the elements within <Instrument> other
than <Microscope> have their own required unique IDs to differentiate
them. All of these elements except <Objective> and <Instrument> can be
referenced from <WavelengthInfo> by their respective IDs.  The
<Instrument> and <Objective> are referenced from <InstrumentRef> within
<Image>.  That's because its understood that the entire image is
captured on one microscope with one objective.

Added <DocumentGroup> element to <OME>.  This allows linking of
multiple OME documents together, for example an entire imaging screen.
Only one of the documents needs to specify the <Instrument>,
<Experimenter>, etc.  The rest of the documents containing one or more
<Image>s need only refer to the <Instrument>s, <Experimenter>s, etc by
their IDs.  This allows the various element IDs to be valid between the
documents in the <DocumentGroup>, instead of only within a single
document.  This of course means that the IDs have to be unique within
the entire group.  The external XML documents are specified by a URI
and a SHA1 digest (for integrity).

<Plate>s are now specified within <OME> rather than <Screen>.  <Plate>s
can belong to one or more screens, which are specified by a list of
<ScreenRef> elements in <Plate>.  This is to normalize the convention
for specifying many-to-many relationships.  The convention is that the
'child' maintains a list of 'parents' that it belongs to.  The 'parent'
does not maintain references to its 'children'.


git-svn-id: svn+ssh://svn.openmicroscopy.org.uk/home/svn/ome/trunk@105 61f9eca8-113b-463c-a1a4-bab9b0a29f14