Corrected logic bug in getImagesInCategory().
The old factory search amounted to:
	$factory->findObjects( 'OME::Image',
		'ClassificationList.Valid'    => 't',
		'ClassificationList.Category' => $category,
		__distinct => 'id',
	);
This worked in most cases, but not all. If an image had been
reclassified, then it has one classification for the old Category where
valid is false, and one classification for the new Cateogry where valid
is true. When getImagesInCategory() was given the old category as a
parameter, this search would find that the image has a classification
for the old category, and it has a valid classification. They just
happen to be separate classifications. Factory searches do not allow
more specific search syntax at this time, so I'm using:
	my @classifications = $factory->findObjects( '@Classification',
		'Valid'    => 't',
		'Category' => $category,
		__order    => 'image.name',
	);
	my @images = map( $_->image, @classifications );


git-svn-id: svn+ssh://svn.openmicroscopy.org.uk/home/svn/ome/trunk@6354 61f9eca8-113b-463c-a1a4-bab9b0a29f14