MAJOR CHANGES. Added flags --force, --compression (no compression is default).
Added warning messages about the size of OMEIS.

The postgress archive is now a text SQL script file. That is supposed to be more portable. I couldn't detect any slow down in comparision with the previous format we were using (compressed custom archive).

I combined extracting OMEmaint and omeDB_backup into one step. When it verifies whether OMEIS is in the archive, it tests for a single file not a whole directory. This should imporve performance.

Function connection() became a part of ome admin configure.


git-svn-id: svn+ssh://svn.openmicroscopy.org.uk/home/svn/ome/trunk@6482 61f9eca8-113b-463c-a1a4-bab9b0a29f14