This resolves a discrepancy between typicv20 calculations and SignatureChainTest calculations. It now uses a 32 bins for the output histogram, and has a default coeeficient value of 11. These values were 40 and 5 before this commit.
typicv20's ChebyFourier uses a hardcoded binning of
	30
SignatureChainTest's ChebyshevFourierTransform uses a hardcoded binning of
	40
Additionally,
typicv20 uses a coefficient value of
	11
SignatureChainTest's ChebyshevFourierTransform disregards the input parameter
and always uses
	5

Analysis:
	The size of a histogram is somewhat arbitrary and approximate performance
can be expected from approximately sized histograms. Use of a 32 bin
histogram is justified for binary data that has a range of 0-2^n where n
>= 5, because it results in evenly sized bins. Additionally, there
should be consistency in the histogram of similar operations. The
version of Chebyshev Statistics included in the merged signatures uses
32 bins.
	Nikita greatly prefers to use a higher coefficient value because he feels
that a low value would disregard the high frequency space, which is important
in some applications.

Resolution:
	Use a 32 bin histogram and a default coefficient value of 11.
ChebyshevFourierTransform has been updated to take this into account, and to
only use the default value when no input parameter is given.


git-svn-id: svn+ssh://svn.openmicroscopy.org.uk/home/svn/ome/trunk@6553 61f9eca8-113b-463c-a1a4-bab9b0a29f14